didn"t like how the postLogin was basically overwriting our authentication logic in passport configs, so i refactored adding a validate function as a seperate middleware, and just calling passport.authenticate() to return the function that uses the logic we defined in passport config
