+++ middleware/auth.js
@@ -1,3 +1,5 @@
+const validator = require("validator");
+
 module.exports = {
   ensureAuth: function (req, res, next) {
     if (req.isAuthenticated()) {
@@ -13,4 +15,22 @@ module.exports = {
       res.redirect("/dashboard");
     }
   },
+  validate: function (req, res, next) {
+    console.log('hey is does this work?', req.body)
+  const validationErrors = [];
+  if (!validator.isEmail(req.body.email))
+    validationErrors.push({ msg: "Please enter a valid email address." });
+  if (validator.isEmpty(req.body.password))
+    validationErrors.push({ msg: "Password cannot be blank." });
+    console.log(validationErrors)
+  if (validationErrors.length) {
+    req.flash("errors", validationErrors);
+    return res.redirect("/login");
+  }
+  req.body.email = validator.normalizeEmail(req.body.email, {
+    gmail_remove_dots: false,
+  });
+  
+  return next();
+  }
 };
